Finish implementation of BankDB class

+++ cpp-in-depth/exercises3/ATM-project/include/Money.h
@@ -8,7 +8,7 @@
 
 class Money
 {
-    unsigned long cents_;
+    unsigned long long cents_;
 
 public:
     // Constructors
